Set up for new way to add client dependencies
Unfortunately, after a fair bit of testing the sed command wasn't playing as nicely as I hoped, so I had to remove the license info from the tops of the files. I also added the initial version of the command I will be using, at least for bash. Windows is going to be a whole new pain. This needs more work though, as it will continue to add new entries every time the script is run in the current state