Added the ability to run the bot on Linux
Using run.sh the bot can now be used (untested) with the server. The callback.js and modules.js will register as invalid due to the way they needed to be formatted for the sed command, but this will not affect program execution. Depending on how batch scripts want to behave, additional tweaks may be required.