Fix some issues introduced during refactoring
Adds esm as a requirement for AVA to remove the requirement for type: module to be in the package.json. Also fixes a couple other bugs. Custom clan list loading and adding/removing clans doesn't work currently as the relative pathing is an issue. Should look to use absolute if possible